fix: resolve GCC sign-conversion warnings in Aesi compareTo integral overload
On Linux (LP64), uint64_t is 'unsigned long' while 'unsigned long long'
is a distinct type. Passing unsigned long long to compareTo(uint64_t)
caused an implicit conversion that GCC flagged as sign-conversion.
Cast through int64_t first (signed widening) then explicitly to uint64_t
to match the parameter type exactly and silence the warning.

+++ include/AesiMultiprecision/Aesi.h
@@ -659,7 +659,7 @@ public:
                 }
             } else if(integral < 0) {
                 if(sign == Negative)
-                    switch(base.compareTo(static_cast<unsigned long long>(-integral))) {
+                    switch(base.compareTo(static_cast<uint64_t>(-static_cast<int64_t>(integral)))) {
                         using enum Comparison;
                         case greater:
                             return less;
@@ -671,7 +671,7 @@ public:
                 else return greater;
             } else {
                 if(sign == Positive)
-                    return base.compareTo(static_cast<unsigned long long>(integral));
+                    return base.compareTo(static_cast<uint64_t>(static_cast<int64_t>(integral)));
                 return less;
             }
         }
